requirements.txt
alembic==1.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.7.2
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
mysql-connector==2.2.9
Pillow==6.1.0
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
redis==3.3.8
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.25.5
visitor==0.1.3
Werkzeug==0.15.6
WTForms==2.2.1
